+++ deepinv/models/__init__.py
@@ -56,5 +56,6 @@ from .bilateral import BilateralFilter
 from .noise_level_estimation import WaveletNoiseEstimator, PatchCovarianceNoiseEstimator
 from .srresnet import SRResNet
 from .ffdnet import FFDNet
+from .wiener import WienerDeconvolution
 
 from .third_party import PromptIR